refactor: Auditing 기능을 슈퍼클래스로 빼서 추상화 진행
