+++ src/main/java/jpa/domain/Station.java
@@ -1,16 +1,10 @@
 package jpa.domain;
 
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
 import javax.persistence.*;
-import java.time.LocalDateTime;
 
 @Entity
 @Table(name = "station")
-@EntityListeners(AuditingEntityListener.class)
-public class Station {
+public class Station extends BaseEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -18,14 +12,6 @@ public class Station {
 
 	@Column(name = "name", nullable = false, unique = true, length = 255)
 	private String name;
-
-	@CreatedDate
-	@Column(name = "created_date", nullable = false, updatable = false, columnDefinition = "datetime")
-	private LocalDateTime createdDate;
-
-	@LastModifiedDate
-	@Column(name = "modified_date", nullable = false, columnDefinition = "datetime")
-	private LocalDateTime modifiedDate;
 
 	public Station() {
 	}
@@ -44,13 +30,5 @@ public class Station {
 
 	void setName(String name) {
 		this.name = name;
-	}
-
-	LocalDateTime getCreatedDate() {
-		return createdDate;
-	}
-
-	LocalDateTime getModifiedDate() {
-		return modifiedDate;
 	}
 }
